More calendar updates
Removed extraeneous classes, fixed styles to work with calendar.
