+++ Workout/.idea/dictionaries/tom_pedraza.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="tom.pedraza">
+    <words>
+      <w>listory</w>
+    </words>
+  </dictionary>
+</component>
