--- Workout/app/src/main/java/com/lostntkdgmail/workout/view/CalendarCustomView.java
@@ -1,110 +1,0 @@
-package com.lostntkdgmail.workout.view;
-
-import android.content.Context;
-import android.util.AttributeSet;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.Button;
-import android.widget.GridView;
-import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.lostntkdgmail.workout.R;
-import com.lostntkdgmail.workout.database.DatabaseAccessor;
-import com.lostntkdgmail.workout.database.LiftTableAccessor;
-
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.List;
-import java.util.Locale;
-public class CalendarCustomView extends LinearLayout{
-    private static final String TAG = CalendarCustomView.class.getSimpleName();
-    private ImageView previousButton, nextButton;
-    private TextView currentDate;
-    private GridView calendarGridView;
-    private Button addEventButton;
-    private static final int MAX_CALENDAR_COLUMN = 42;
-    private int month, year;
-    private SimpleDateFormat formatter = new SimpleDateFormat("MMMM yyyy", Locale.ENGLISH);
-    private Calendar cal = Calendar.getInstance(Locale.ENGLISH);
-    private Context context;
-    private GridAdapter mAdapter;
-    private LiftTableAccessor mQuery;
-    public CalendarCustomView(Context context) {
-        super(context);
-    }
-
-    public CalendarCustomView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        this.context = context;
-        initializeUILayout();
-        setUpCalendarAdapter();
-        setPreviousButtonClickEvent();
-        setNextButtonClickEvent();
-        setGridCellClickEvents();
-        Log.d(TAG, "I need to call this method");
-    }
-
-    public CalendarCustomView(Context context, AttributeSet attrs, int defStyleAttr) {
-        super(context, attrs, defStyleAttr);
-    }
-    private void initializeUILayout(){
-        LayoutInflater inflater = (LayoutInflater) context.getSystemService(context.LAYOUT_INFLATER_SERVICE);
-        View view = inflater.inflate(R.layout.calendar_view, this);
-        previousButton = (ImageView)view.findViewById(R.id.previous_month);
-        nextButton = (ImageView)view.findViewById(R.id.next_month);
-        currentDate = (TextView)view.findViewById(R.id.display_current_date);
-        addEventButton = (Button)view.findViewById(R.id.add_calendar_event);
-        calendarGridView = (GridView)view.findViewById(R.id.calendar_grid);
-    }
-    private void setPreviousButtonClickEvent(){
-        previousButton.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                cal.add(Calendar.MONTH, -1);
-                setUpCalendarAdapter();
-            }
-        });
-    }
-    private void setNextButtonClickEvent(){
-        nextButton.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                cal.add(Calendar.MONTH, 1);
-                setUpCalendarAdapter();
-            }
-        });
-    }
-    private void setGridCellClickEvents(){
-        calendarGridView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Toast.makeText(context, "Clicked " + position, Toast.LENGTH_LONG).show();
-            }
-        });
-    }
-    private void setUpCalendarAdapter(){
-        List<Date> dayValueInCells = new ArrayList<Date>();
-        mQuery = new LiftTableAccessor(getContext());
-        List<EventObjects> mEvents = mQuery.getAllLifts();
-        Calendar mCal = (Calendar)cal.clone();
-        mCal.set(Calendar.DAY_OF_MONTH, 1);
-        int firstDayOfTheMonth = mCal.get(Calendar.DAY_OF_WEEK) - 1;
-        mCal.add(Calendar.DAY_OF_MONTH, -firstDayOfTheMonth);
-        while(dayValueInCells.size() < MAX_CALENDAR_COLUMN){
-            dayValueInCells.add(mCal.getTime());
-            mCal.add(Calendar.DAY_OF_MONTH, 1);
-        }
-        Log.d(TAG, "Number of date " + dayValueInCells.size());
-        String sDate = formatter.format(cal.getTime());
-        currentDate.setText(sDate);
-        mAdapter = new GridAdapter(context, dayValueInCells, cal, mEvents);
-        calendarGridView.setAdapter(mAdapter);
-    }
-}
