+++ Workout/app/src/main/java/com/lostntkdgmail/workout/view/GridAdapter.java
@@ -3,8 +3,6 @@ package com.lostntkdgmail.workout.view;
 import android.content.Context;
 
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -17,6 +15,9 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 public class GridAdapter extends ArrayAdapter {
+
+    //TODO increase the height of each box
+    //TODO pass the full list of lifts here so it can be marked by the calendar
     private static final String TAG = GridAdapter.class.getSimpleName();
     private LayoutInflater mInflater;
     private List<Date> monthlyDates;
@@ -29,7 +30,7 @@ public class GridAdapter extends ArrayAdapter {
         this.allEvents = allEvents;
         mInflater = LayoutInflater.from(context);
     }
-    @NonNull
+
     @Override
     public View getView(int position, View convertView, ViewGroup parent) {
         Date mDate = monthlyDates.get(position);
@@ -45,9 +46,9 @@ public class GridAdapter extends ArrayAdapter {
             view = mInflater.inflate(R.layout.single_cell_layout, parent, false);
         }
         if(displayMonth == currentMonth && displayYear == currentYear){
-            view.setBackgroundColor(Color.parseColor("#FF5733"));
+            view.setBackgroundColor(Color.parseColor("#000000"));
         }else{
-            view.setBackgroundColor(Color.parseColor("#cccccc"));
+            view.setBackgroundColor(Color.parseColor("#373737"));
         }
         //Add day to calendar
         TextView cellNumber = (TextView)view.findViewById(R.id.calendar_date_id);
@@ -68,11 +69,12 @@ public class GridAdapter extends ArrayAdapter {
     public int getCount() {
         return monthlyDates.size();
     }
-    @Nullable
+
     @Override
     public Object getItem(int position) {
         return monthlyDates.get(position);
     }
+
     @Override
     public int getPosition(Object item) {
         return monthlyDates.indexOf(item);
