+++ Workout/app/src/main/res/layout/activity_main.xml
@@ -5,8 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.MainActivity"
-    android:background="@color/colorPrimary">
+    tools:context=".main.MainActivity">
 
     <com.lostntkdgmail.workout.main.NonSwipeViewPager
         android:id="@+id/container"
