--- Workout/app/src/main/res/layout/calendar_fragment_view.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_custom_calendar"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-    <com.inducesmile.androidcustomcalendar.CalendarCustomView
-        android:id="@+id/custom_calendar"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-    </com.inducesmile.androidcustomcalendar.CalendarCustomView>
-</LinearLayout>
