+++ Workout/app/src/main/res/layout/calendar_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.lostntkdgmail.workout.view.CalendarCustomView xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/activity_custom_calendar"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -24,7 +24,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:gravity="center"
-            android:textColor="@color/colorPrimaryDark"
+            android:textColor="@color/textPrimary"
             android:text="@string/current_date"
             android:layout_marginTop="4dp"
             android:textStyle="bold"
@@ -33,23 +33,14 @@
             android:id="@+id/next_month"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_gravity="left"
+            android:layout_gravity="start"
             android:contentDescription="@string/app_name"
             android:src="@drawable/arrow_right"
             android:layout_weight="1"/>
     </LinearLayout>
-    <Button
-        android:id="@+id/add_calendar_event"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textStyle="bold"
-        android:textSize="12sp"
-        android:text="@string/add_event"
-        android:gravity="center"/>
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/colorPrimaryDark"
         android:paddingBottom="4dp"
         android:paddingTop="4dp"
         android:orientation="horizontal">
@@ -116,7 +107,7 @@
         <TextView
             android:id="@+id/sat"
             android:layout_width="0dp"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:gravity="center"
             android:textColor="@color/colorWhite"
             android:text="@string/sat"
@@ -129,4 +120,4 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:numColumns="7"/>
-</com.lostntkdgmail.workout.view.CalendarCustomView>
+</LinearLayout>
