+++ Workout/app/src/main/res/values/colors.xml
@@ -8,4 +8,9 @@
     <color name="colorAccent">#373737</color>
     <color name="colorWhite">#ffffff</color>
     <color name="colorBlack">#000000</color>
+    <color name="primary_dark_material_dark">#000000</color>
+    <color name="foreground_material_dark">#313131</color>
+    <color name="background_material_dark">#000000</color>
+    <color name="text_color_primary">#ffffff</color>
+    <color name="primary_material_dark">#000000</color>
 </resources>
