+++ Workout/app/src/main/res/values/styles.xml
@@ -1,8 +1,9 @@
 <resources>
 
-    <style name="AppTheme" parent="Theme.AppCompat.NoActionBar">
-        <item name="android:textColorPrimary">@color/textPrimary</item>
-        <item name="android:background">@color/colorPrimary</item>
+    <style name="fuckThemes" parent="Theme.AppCompat" >
+        <item name="android:colorBackground">@color/background_material_dark</item>
+        <item name="android:colorForeground">@color/foreground_material_dark</item>
+        <item name="android:textColorPrimary">@color/text_color_primary</item>
     </style>
